render >1 workspace at a time
